vstsdk370_30_07_2020_build_115

+++ source/vst/vstpresetfile.cpp
@@ -8,7 +8,7 @@
 //
 //-----------------------------------------------------------------------------
 // LICENSE
-// (c) 2019, Steinberg Media Technologies GmbH, All Rights Reserved
+// (c) 2020, Steinberg Media Technologies GmbH, All Rights Reserved
 //-----------------------------------------------------------------------------
 // Redistribution and use in source and binary forms, with or without modification,
 // are permitted provided that the following conditions are met:
@@ -863,7 +863,7 @@ BufferStream::~BufferStream () {FUNKNOWN_DTOR}
 //------------------------------------------------------------------------
 tresult PLUGIN_API BufferStream::read (void* buffer, int32 numBytes, int32* numBytesRead)
 {
-	uint32 size = mBuffer.get (buffer, static_cast<size_t> (numBytes));
+	uint32 size = mBuffer.get (buffer, static_cast<uint32> (numBytes));
 	if (numBytesRead)
 		*numBytesRead = static_cast<int32> (size);
 
@@ -873,7 +873,7 @@ tresult PLUGIN_API BufferStream::read (void* buffer, int32 numBytes, int32* numB
 //------------------------------------------------------------------------
 tresult PLUGIN_API BufferStream::write (void* buffer, int32 numBytes, int32* numBytesWritten)
 {
-	bool res = mBuffer.put (buffer, static_cast<size_t> (numBytes));
+	bool res = mBuffer.put (buffer, static_cast<uint32> (numBytes));
 	if (numBytesWritten)
 		*numBytesWritten = res ? numBytes : 0;
